Changed the "Holo interface" to "AppCompact", is similar to HOLO. We need use this to made compatible with android api level 10 (Android 2.3.3).
This is the official solution of Google/Android.

We have some changes to use "AppCompact" in the code, and we have other changes too...

I've changed the Main icons (search, barcode) and make the search extensible button.

I  put the "Not implemented" text/dialogs.

Note:
Sorry 4 my English, I'm spanish a the app development is in Spanish, but i make the comments in "my bad english".
Thanks